requirements.txt
pdfplumber
pytesseract
pdf2image
Pillow
PyMuPDF
python-docx
tqdm
dateparser
flashtext  
transformers
sentence-transformers
nomic
sqlite3
requests
pandas
faiss-cpu
torch
jsonlines
tqdm
streamlit
spacy
python-dotenv
langchain
torch 
phonenumbers
requests 
beautifulsoup4